Create Docker
Here's a breakdown of what each step does:

We use the official Nginx base image to build our container.
Set the working directory inside the container to /usr/share/nginx/html, which is the default Nginx document root directory.

We create a new user named "myuser" using the useradd command, and set /bin/bash as the default shell and /home/myuser as the home directory.
We copy the website files from the current directory (where the Dockerfile is located) to the container's document root (/usr/share/nginx/html).
We change the ownership of the copied files to the new user "myuser" using chown.

We switch to the user "myuser" so that Nginx runs under this non-privileged user.
Expose port 80, which is the default port for Nginx to serve web content.
Finally, we start Nginx using the CMD directive with the necessary arguments to run it as the main process in the container.
To build the Docker image, save the above Dockerfile in a directory and then run the following command from that directory:


docker build -t my_nginx_image .

This will create a Docker image named my_nginx_image with the Nginx server and your website files inside it. To run a container using this image, execute the following command:


docker run -d -p 80:80 my_nginx_image

This will start the container in the background, mapping the container's port 80 to the host's port 80, allowing you to access the Nginx server and the website files from your web browser at http://localhost.